[FE] refactor: 커스텀 훅 네이밍 변경
- 커스텀 훅 네이밍을 변경

- issue (#109)
