+++ client/src/hooks/index.ts
@@ -1,3 +1,3 @@
 export * from './store';
-export * from './auth';
-export * from './util';
+export * from './useSilentRefresh';
+export * from './useLoader';
