+++ client/src/hooks/useSilentRefresh.ts
